Refactor circle and rounded corners transformations to share a common paint.
