+++ coil-core/src/androidMain/kotlin/coil3/transform/RoundedCornersTransformation.kt
@@ -1,14 +1,8 @@
 package coil3.transform
 
 import android.graphics.Bitmap
-import android.graphics.BitmapShader
-import android.graphics.Color
-import android.graphics.Matrix
-import android.graphics.Paint
 import android.graphics.Path
-import android.graphics.PorterDuff
 import android.graphics.RectF
-import android.graphics.Shader
 import androidx.annotation.Px
 import androidx.core.graphics.applyCanvas
 import androidx.core.graphics.createBitmap
@@ -54,43 +48,23 @@ class RoundedCornersTransformation(
     override val cacheKey = "${this::class.qualifiedName}-$topLeft,$topRight,$bottomLeft,$bottomRight"
 
     override suspend fun transform(input: Bitmap, size: Size): Bitmap {
-        val paint = Paint(Paint.ANTI_ALIAS_FLAG or Paint.FILTER_BITMAP_FLAG)
-
         val (outputWidth, outputHeight) = calculateOutputSize(input, size)
-
-        val output = createBitmap(outputWidth, outputHeight, input.safeConfig)
-        output.applyCanvas {
-            drawColor(Color.TRANSPARENT, PorterDuff.Mode.CLEAR)
-
-            val matrix = Matrix()
-            val multiplier = DecodeUtils.computeSizeMultiplier(
-                srcWidth = input.width,
-                srcHeight = input.height,
-                dstWidth = outputWidth,
-                dstHeight = outputHeight,
-                scale = Scale.FILL,
-            ).toFloat()
-            val dx = (outputWidth - multiplier * input.width) / 2
-            val dy = (outputHeight - multiplier * input.height) / 2
-            matrix.setTranslate(dx, dy)
-            matrix.preScale(multiplier, multiplier)
-
-            val shader = BitmapShader(input, Shader.TileMode.CLAMP, Shader.TileMode.CLAMP)
-            shader.setLocalMatrix(matrix)
-            paint.shader = shader
-
-            val radii = floatArrayOf(
-                topLeft, topLeft,
-                topRight, topRight,
-                bottomRight, bottomRight,
-                bottomLeft, bottomLeft,
-            )
-            val rect = RectF(0f, 0f, width.toFloat(), height.toFloat())
-            val path = Path().apply { addRoundRect(rect, radii, Path.Direction.CW) }
-            drawPath(path, paint)
+        return createBitmap(outputWidth, outputHeight, input.safeConfig).applyCanvas {
+            val paint = newBitmapShaderPaint(input, outputWidth, outputHeight)
+            if (topLeft == topRight && topRight == bottomLeft && bottomLeft == bottomRight) {
+                drawRoundRect(0f, 0f, outputWidth.toFloat(), outputHeight.toFloat(), topLeft, topLeft, paint)
+            } else {
+                val radii = floatArrayOf(
+                    topLeft, topLeft,
+                    topRight, topRight,
+                    bottomRight, bottomRight,
+                    bottomLeft, bottomLeft,
+                )
+                val rect = RectF(0f, 0f, outputWidth.toFloat(), outputHeight.toFloat())
+                val path = Path().apply { addRoundRect(rect, radii, Path.Direction.CW) }
+                drawPath(path, paint)
+            }
         }
-
-        return output
     }
 
     private fun calculateOutputSize(input: Bitmap, size: Size): IntPair {
